# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OCR Information Extraction API - Schema-anchored extraction pipeline
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

/**
 * Data type for this field
 */
export type SchemaFieldType =
  (typeof SchemaFieldType)[keyof typeof SchemaFieldType];

export const SchemaFieldType = {
  string: "string",
  number: "number",
  date: "date",
  boolean: "boolean",
  array: "array",
} as const;

export interface SchemaField {
  /** Field identifier (snake_case) */
  name: string;
  /** Human-readable field label */
  label: string;
  /** Data type for this field */
  type: SchemaFieldType;
  /** What this field represents and how to extract it */
  description: string;
  /** Whether this field must be present in every extraction */
  required: boolean;
  /** Example value for this field */
  example?: string;
}

export interface DocumentSchema {
  id: number;
  /** Schema name (e.g. "Receipt", "Invoice") */
  name: string;
  /** What type of document this schema is for */
  description: string;
  fields: SchemaField[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateSchemaBody {
  name: string;
  description: string;
  fields: SchemaField[];
}

export interface ExtractedField {
  name: string;
  label: string;
  /** Extracted value (null if not found) */
  value?: unknown;
  /** Confidence score 0-1 */
  confidence: number;
  /** Whether the field was found in the document */
  present: boolean;
}

export type ExtractionResultStatus =
  (typeof ExtractionResultStatus)[keyof typeof ExtractionResultStatus];

export const ExtractionResultStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface ExtractionResult {
  id: number;
  schemaId: number;
  schemaName: string;
  fileName: string;
  fileType: string;
  status: ExtractionResultStatus;
  fields: ExtractedField[];
  /** Average confidence across all fields */
  overallConfidence: number;
  processingTimeMs: number;
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ListSchemas200 = {
  schemas: DocumentSchema[];
};

export type DeleteSchema200 = {
  success: boolean;
};

export type ListExtractionsParams = {
  schemaId?: number;
};

export type ListExtractions200 = {
  extractions: ExtractionResult[];
};

export type GetExtractionRawText200 = {
  rawText: string;
};