# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OCR Information Extraction API - Schema-anchored extraction pipeline
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all defined extraction schemas
 * @summary List document schemas
 */
export const ListSchemasResponse = zod.object({
  schemas: zod.array(
    zod.object({
      id: zod.number(),
      name: zod
        .string()
        .describe('Schema name (e.g. \"Receipt\", \"Invoice\")'),
      description: zod
        .string()
        .describe("What type of document this schema is for"),
      fields: zod.array(
        zod.object({
          name: zod.string().describe("Field identifier (snake_case)"),
          label: zod.string().describe("Human-readable field label"),
          type: zod
            .enum(["string", "number", "date", "boolean", "array"])
            .describe("Data type for this field"),
          description: zod
            .string()
            .describe("What this field represents and how to extract it"),
          required: zod
            .boolean()
            .describe("Whether this field must be present in every extraction"),
          example: zod
            .string()
            .optional()
            .describe("Example value for this field"),
        }),
      ),
      createdAt: zod.date(),
      updatedAt: zod.date(),
    }),
  ),
});

/**
 * Define a new extraction schema with fields
 * @summary Create a document schema
 */
export const CreateSchemaBody = zod.object({
  name: zod.string(),
  description: zod.string(),
  fields: zod.array(
    zod.object({
      name: zod.string().describe("Field identifier (snake_case)"),
      label: zod.string().describe("Human-readable field label"),
      type: zod
        .enum(["string", "number", "date", "boolean", "array"])
        .describe("Data type for this field"),
      description: zod
        .string()
        .describe("What this field represents and how to extract it"),
      required: zod
        .boolean()
        .describe("Whether this field must be present in every extraction"),
      example: zod.string().optional().describe("Example value for this field"),
    }),
  ),
});

/**
 * @summary Get a document schema
 */
export const GetSchemaParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSchemaResponse = zod.object({
  id: zod.number(),
  name: zod.string().describe('Schema name (e.g. \"Receipt\", \"Invoice\")'),
  description: zod
    .string()
    .describe("What type of document this schema is for"),
  fields: zod.array(
    zod.object({
      name: zod.string().describe("Field identifier (snake_case)"),
      label: zod.string().describe("Human-readable field label"),
      type: zod
        .enum(["string", "number", "date", "boolean", "array"])
        .describe("Data type for this field"),
      description: zod
        .string()
        .describe("What this field represents and how to extract it"),
      required: zod
        .boolean()
        .describe("Whether this field must be present in every extraction"),
      example: zod.string().optional().describe("Example value for this field"),
    }),
  ),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a document schema
 */
export const DeleteSchemaParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteSchemaResponse = zod.object({
  success: zod.boolean(),
});

/**
 * Returns all past extraction results
 * @summary List all extractions
 */
export const ListExtractionsQueryParams = zod.object({
  schemaId: zod.coerce.number().optional(),
});

export const ListExtractionsResponse = zod.object({
  extractions: zod.array(
    zod.object({
      id: zod.number(),
      schemaId: zod.number(),
      schemaName: zod.string(),
      fileName: zod.string(),
      fileType: zod.string(),
      status: zod.enum(["pending", "processing", "completed", "failed"]),
      fields: zod.array(
        zod.object({
          name: zod.string(),
          label: zod.string(),
          value: zod
            .unknown()
            .optional()
            .describe("Extracted value (null if not found)"),
          confidence: zod.number().describe("Confidence score 0-1"),
          present: zod
            .boolean()
            .describe("Whether the field was found in the document"),
        }),
      ),
      overallConfidence: zod
        .number()
        .describe("Average confidence across all fields"),
      processingTimeMs: zod.number(),
      errorMessage: zod.string().nullish(),
      createdAt: zod.date(),
      updatedAt: zod.date(),
    }),
  ),
});

/**
 * @summary Get an extraction result
 */
export const GetExtractionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetExtractionResponse = zod.object({
  id: zod.number(),
  schemaId: zod.number(),
  schemaName: zod.string(),
  fileName: zod.string(),
  fileType: zod.string(),
  status: zod.enum(["pending", "processing", "completed", "failed"]),
  fields: zod.array(
    zod.object({
      name: zod.string(),
      label: zod.string(),
      value: zod
        .unknown()
        .optional()
        .describe("Extracted value (null if not found)"),
      confidence: zod.number().describe("Confidence score 0-1"),
      present: zod
        .boolean()
        .describe("Whether the field was found in the document"),
    }),
  ),
  overallConfidence: zod
    .number()
    .describe("Average confidence across all fields"),
  processingTimeMs: zod.number(),
  errorMessage: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Get raw OCR text for an extraction
 */
export const GetExtractionRawTextParams = zod.object({
  id: zod.coerce.number(),
});

export const GetExtractionRawTextResponse = zod.object({
  rawText: zod.string(),
});