# lib/api-zod/src/generated/types/createSchemaBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OCR Information Extraction API - Schema-anchored extraction pipeline
 * OpenAPI spec version: 0.1.0
 */
import type { SchemaField } from "./schemaField";

export interface CreateSchemaBody {
  name: string;
  description: string;
  fields: SchemaField[];
}