# lib/api-zod/src/generated/types/documentSchema.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OCR Information Extraction API - Schema-anchored extraction pipeline
 * OpenAPI spec version: 0.1.0
 */
import type { SchemaField } from "./schemaField";

export interface DocumentSchema {
  id: number;
  /** Schema name (e.g. "Receipt", "Invoice") */
  name: string;
  /** What type of document this schema is for */
  description: string;
  fields: SchemaField[];
  createdAt: Date;
  updatedAt: Date;
}