# lib/api-zod/src/generated/types/extractedField.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OCR Information Extraction API - Schema-anchored extraction pipeline
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedField {
  name: string;
  label: string;
  /** Extracted value (null if not found) */
  value?: unknown;
  /** Confidence score 0-1 */
  confidence: number;
  /** Whether the field was found in the document */
  present: boolean;
}