# lib/api-zod/src/generated/types/extractionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OCR Information Extraction API - Schema-anchored extraction pipeline
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedField } from "./extractedField";
import type { ExtractionResultStatus } from "./extractionResultStatus";

export interface ExtractionResult {
  id: number;
  schemaId: number;
  schemaName: string;
  fileName: string;
  fileType: string;
  status: ExtractionResultStatus;
  fields: ExtractedField[];
  /** Average confidence across all fields */
  overallConfidence: number;
  processingTimeMs: number;
  errorMessage?: string | null;
  createdAt: Date;
  updatedAt: Date;
}