# lib/api-zod/src/generated/types/schemaField.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OCR Information Extraction API - Schema-anchored extraction pipeline
 * OpenAPI spec version: 0.1.0
 */
import type { SchemaFieldType } from "./schemaFieldType";

export interface SchemaField {
  /** Field identifier (snake_case) */
  name: string;
  /** Human-readable field label */
  label: string;
  /** Data type for this field */
  type: SchemaFieldType;
  /** What this field represents and how to extract it */
  description: string;
  /** Whether this field must be present in every extraction */
  required: boolean;
  /** Example value for this field */
  example?: string;
}